test: merge serial tests in ddl, infoschema, session, store, table, telementry and types
