--- ddl/db_cache_serial_test.go
@@ -1,97 +1,0 @@
-// Copyright 2021 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package ddl_test
-
-import (
-	"testing"
-	"time"
-
-	"github.com/pingcap/tidb/ddl"
-	"github.com/pingcap/tidb/domain"
-	"github.com/pingcap/tidb/errno"
-	"github.com/pingcap/tidb/parser/model"
-	"github.com/pingcap/tidb/parser/terror"
-	"github.com/pingcap/tidb/session"
-	"github.com/pingcap/tidb/store/mockstore"
-	"github.com/pingcap/tidb/testkit"
-	"github.com/stretchr/testify/require"
-)
-
-func TestAlterTableCache(t *testing.T) {
-	store, err := mockstore.NewMockStore()
-	require.NoError(t, err)
-	session.SetSchemaLease(600 * time.Millisecond)
-	session.DisableStats4Test()
-	dom, err := session.BootstrapSession(store)
-	require.NoError(t, err)
-
-	dom.SetStatsUpdating(true)
-
-	clean := func() {
-		dom.Close()
-		err := store.Close()
-		require.NoError(t, err)
-	}
-	defer clean()
-	tk := testkit.NewTestKit(t, store)
-	tk2 := testkit.NewTestKit(t, store)
-
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t1")
-	tk2.MustExec("use test")
-	/* Test of cache table */
-	tk.MustExec("create table t1 ( n int auto_increment primary key)")
-	tk.MustGetErrCode("alter table t1 ca", errno.ErrParse)
-	tk.MustGetErrCode("alter table t2 cache", errno.ErrNoSuchTable)
-	tk.MustExec("alter table t1 cache")
-	checkTableCacheStatus(t, tk.Session(), "test", "t1", model.TableCacheStatusEnable)
-	tk.MustExec("drop table if exists t1")
-	/*Test can't skip schema checker*/
-	tk.MustExec("drop table if exists t1,t2")
-	tk.MustExec("CREATE TABLE t1 (a int)")
-	tk.MustExec("CREATE TABLE t2 (a int)")
-	tk.MustExec("begin")
-	tk.MustExec("insert into t1 set a=1;")
-	tk2.MustExec("alter table t1 cache;")
-	_, err = tk.Exec("commit")
-	require.True(t, terror.ErrorEqual(domain.ErrInfoSchemaChanged, err))
-	/* Test can skip schema checker */
-	tk.MustExec("begin")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("CREATE TABLE t1 (a int)")
-	tk.MustExec("insert into t1 set a=2;")
-	tk2.MustExec("alter table t2 cache")
-	tk.MustExec("commit")
-	// Test if a table is not exists
-	tk.MustExec("drop table if exists t")
-	tk.MustGetErrCode("alter table t cache", errno.ErrNoSuchTable)
-	tk.MustExec("create table t (a int)")
-	tk.MustExec("alter table t cache")
-	// Multiple alter cache is okay
-	tk.MustExec("alter table t cache")
-	tk.MustExec("alter table t cache")
-	// Test a temporary table
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create temporary table t (id int primary key auto_increment, u int unique, v int)")
-	tk.MustExec("drop table if exists tmp1")
-	// local temporary table alter is not supported
-	tk.MustGetErrCode("alter table t cache", errno.ErrUnsupportedDDLOperation)
-	// test global temporary table
-	tk.MustExec("create global temporary table tmp1 " +
-		"(id int not null primary key, code int not null, value int default null, unique key code(code))" +
-		"on commit delete rows")
-	tk.MustGetErrMsg("alter table tmp1 cache", ddl.ErrOptOnTemporaryTable.GenWithStackByArgs("alter temporary table cache").Error())
-
-}
