+++ store/driver/config_test.go
