+++ store/driver/snap_interceptor_test.go
