+++ telemetry/telemetry_test.go
@@ -101,3 +101,29 @@ func TestPreview(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, "", r)
 }
+
+func TestReport(t *testing.T) {
+	if runtime.GOOS == "windows" {
+		t.Skip("integration.NewClusterV3 will create file contains a colon which is not allowed on Windows")
+	}
+
+	etcdCluster := integration.NewClusterV3(t, &integration.ClusterConfig{Size: 1})
+	defer etcdCluster.Terminate(t)
+	store, clean := testkit.CreateMockStore(t)
+	defer clean()
+	se, err := session.CreateSession4Test(store)
+	require.NoError(t, err)
+	defer se.Close()
+
+	config.GetGlobalConfig().EnableTelemetry = false
+	require.NoError(t, telemetry.ReportUsageData(se, etcdCluster.RandClient()))
+
+	status, err := telemetry.GetTelemetryStatus(etcdCluster.RandClient())
+	require.NoError(t, err)
+
+	jsonParsed, err := gabs.ParseJSON([]byte(status))
+	require.NoError(t, err)
+	require.True(t, jsonParsed.Path("is_error").Data().(bool))
+	require.Equal(t, "telemetry is disabled", jsonParsed.Path("error_msg").Data().(string))
+	require.False(t, jsonParsed.Path("is_request_sent").Data().(bool))
+}
